Recover wsrep position even if wsrep is off
Before this commit, starting server with --wsrep-recovery
returned random position from uninitialized memory. This was
because the position was not read from storage engine and
the out parameters were not set in wsrep_get_SE_checkpoint().

In order to fix this, removed the check for WSREP_ON at the
beginning of wsrep_get_SE_checkpoint() to run the recovery
regardless if wsrep_on is ON or OFF.

Also fixed casting error in trx_sys_read_wsrep_checkpoint()
to read the stored formatID=-1 as -1 instead of UINT_MAX.
This was required to avoid logging 'Read non-wsrep XID from
storage engines' warning.

+++ storage/innobase/trx/trx0sys.cc
@@ -380,18 +380,15 @@ trx_sys_read_wsrep_checkpoint(XID* xid)
                 return;
         }
 
-        xid->set_format_id((long)mach_read_from_4(
+        xid->set_format_id((int)mach_read_from_4(
                 sys_header
                 + TRX_SYS_WSREP_XID_INFO + TRX_SYS_WSREP_XID_FORMAT));
-        xid->set_gtrid_length((long)mach_read_from_4(
+        xid->set_gtrid_length((int)mach_read_from_4(
                 sys_header
                 + TRX_SYS_WSREP_XID_INFO + TRX_SYS_WSREP_XID_GTRID_LEN));
-        xid->set_bqual_length((long)mach_read_from_4(
+        xid->set_bqual_length((int)mach_read_from_4(
                 sys_header
                 + TRX_SYS_WSREP_XID_INFO + TRX_SYS_WSREP_XID_BQUAL_LEN));
-        //ut_memcpy(xid->data,
-        //          sys_header + TRX_SYS_WSREP_XID_INFO + TRX_SYS_WSREP_XID_DATA,
-        //          XIDDATASIZE);
         xid->set_data(
                       sys_header + TRX_SYS_WSREP_XID_INFO + TRX_SYS_WSREP_XID_DATA, 
                       XIDDATASIZE);
